test: Updated utils file to move inside dashboards
