+++ tests/ui-testing/playwright-tests/dashboards/custom-charts.spec.js
@@ -1,8 +1,8 @@
 import { test, expect } from "../baseFixtures";
 import path from "path";
 import fs from "fs";
-import { login } from "../utils/dashLogin";
-import { ingestion } from "../utils/dashIngestion";
+import { login } from "./utils/dashLogin.js";
+import { ingestion } from "./utils/dashIngestion.js";
 import DashboardPanel from "../../pages/dashboardPages/dashboard-panel-edit";
 import { DashboardPage } from "../../pages/dashboardPage";
 import DashboardactionPage from "../../pages/dashboardPages/dashboard-panel-actions";
